GEOS-1225, changed default camera refresh + kmscore
git-svn-id: http://svn.codehaus.org/geoserver/trunk@7298 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 91c56cf6bea5c5788917a6c0b3afd44a37e3a652 [formerly ce296010f9d3fcde4d2342b976c50ea7a532fe93]
Former-commit-id: 5f8fb6e91cdfeb118f51e0b61001d557691bd069

+++ geoserver/wms/src/main/java/org/vfny/geoserver/wms/responses/map/kml/KMLReflector.java
@@ -29,7 +29,7 @@ public class KMLReflector {
     public static final String KMZ_MIME_TYPE = "application/vnd.google-earth.kmz+xml";
 
     /** default 'kmscore' value */
-    public static final Integer KMSCORE = new Integer(30);
+    public static final Integer KMSCORE = new Integer(50);
 
     /** default 'kmattr' value */
     public static final Boolean KMATTR = Boolean.TRUE;
